+++ src/editor/plugins/lsp/kcl/parser.ts
@@ -9,8 +9,8 @@ import {
   NodeType,
   NodeSet,
 } from '@lezer/common'
-import { LanguageServerClient } from '.'
-import { posToOffset } from 'editor/lsp/util'
+import { LanguageServerClient } from 'editor/plugins/lsp'
+import { posToOffset } from 'editor/plugins/lsp/util'
 import { SemanticToken } from './semantic_tokens'
 import { DocInput } from '@codemirror/language'
 import { tags, styleTags } from '@lezer/highlight'
