+++ src/editor/plugins/lsp/server-capability-registration.ts
